add rate limiter for user signup, login and forgotpassword apis, add condition for name must be string in signup, create getUser by id api, update enum of user's role and set timestamps: true, restrict getAllUsers api for admin only, create error handler for schema validation and casterror in global error handler, create category model and apis for create, get, getall and update category

+++ backend/src/controllers/user.controller.js
@@ -14,9 +14,19 @@ const filterFields = (obj, ...allfields) => {
 const getAllUsers = catchAsync(async(req, res, next) => {
     const allUsers = await User.find();
 
-    if(!allUsers) return next( new AppError("user not found", 404));
+    if(allUsers.length === 0) return next( new AppError("user not found", 404));
 
     res.status(200).json({status: "success", data: allUsers});
+})
+
+const getUser =  catchAsync(async (req, res, next) => {
+    const userId = req.params.id;
+    if(!userId) return next(new AppError("userId not found", 404))
+
+    const user = await User.findById(userId);
+    if(!user) return next( new AppError("user not found", 404));
+
+    res.status(200).json({status: "success", data: user});
 })
 
 const updateMe = catchAsync( async(req, res, next) => {
@@ -29,22 +39,24 @@ const updateMe = catchAsync( async(req, res, next) => {
     const filteredBody = filterFields(req.body, "name", "email");
     
     // user allowed to change only name & email
-    const updatedUser = await User.findByIdAndUpdate(
+    const user = await User.findByIdAndUpdate(
         req.user.id, 
         filteredBody, 
         {new: true, runValidators: true}
     )
-    res.status(200).json({status: "success", updatedUser: updatedUser})
+    res.status(200).json({status: "success", updatedUser: user})
 })
 
 const deleteMe = catchAsync(async (req, res, next) => {
+    const user = await User.findByIdAndUpdate(req.user.id, {active: false});
+    if(!user) return next(new AppError("User not found", 404));
 
-    await User.findByIdAndUpdate(req.user.id, {active: false})
     res.status(204).json({status: "success", data: null});
 })
 
 const deleteUser = catchAsync(async (req, res, next) => {
     const id = req.params.id;
+    if(!id) return next("id not found", 404)
 
     const deletedUser = await User.findByIdAndDelete(id);
     if(!deletedUser) return next( new AppError("user not found", 404));
@@ -52,4 +64,4 @@ const deleteUser = catchAsync(async (req, res, next) => {
     res.status(200).json({status: "success", deletedUser: deletedUser});
 })
 
-module.exports = {getAllUsers, updateMe, deleteMe, deleteUser}
+module.exports = {getAllUsers, getUser, updateMe, deleteMe, deleteUser}
